+++ arduino/ambovis/menu.h
@@ -44,7 +44,7 @@ class Menu_inic {
   bool switching_menus;
   int m_curr_sel;
   public:
-    Menu_inic(byte *mode, byte *bpm, byte *i_e);
+    Menu_inic(byte *mode, short *bpm, byte *i_e);
     void clear_n_sel( int menu );
     void Menu_inic::check_encoder ( );
     void Menu_inic::display_lcd ( );
